displaying reviewer image , count of reviews for specific product

+++ recommendation/src/main/java/com/stackroute/recommendation/service/ProductServiceImpl.java
@@ -3,6 +3,7 @@ package com.stackroute.recommendation.service;
 import com.stackroute.recommendation.domain.Product;
 import com.stackroute.recommendation.dto.ProductDTO;
 import com.stackroute.recommendation.dto.ReviewDTO;
+import com.stackroute.recommendation.exception.ProductNotFoundException;
 import com.stackroute.recommendation.repository.ProductRepository;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,8 +25,14 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public Collection<Product> getAll() {
-        return productRepository.getAllProducts();
+    public Collection<Product> getAll()throws Exception {
+        Collection<Product> productCollection=productRepository.getAllProducts();
+            if(productCollection.isEmpty()){
+                throw  new Exception("Products not available");
+            }
+        else {
+                return productRepository.getAllProducts();
+            }
     }
 
     @Override
@@ -37,20 +44,38 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public Collection<Product> getByFamily(String productFamily) {
-
-        return productRepository.getNode(productFamily);
+    public Collection<Product> getByFamily(String productFamily) throws ProductNotFoundException {
+       Collection<Product> productCollection=productRepository.getNode(productFamily);
+           if(productCollection.isEmpty()){
+               throw  new ProductNotFoundException("Products not available");
+           }
+           else {
+               return productRepository.getNode(productFamily);
+           }
     }
 
     @Override
-    public Collection<Product> getBySubCategory(String subCategory) {
-        return productRepository.getBysubCategory(subCategory);
+    public Collection<Product> getBySubCategory(String subCategory) throws ProductNotFoundException {
+        Collection<Product> productCollection=productRepository.getBysubCategory(subCategory);
+        if(productCollection.isEmpty()){
+            throw  new ProductNotFoundException("Products not available");
+        }
+        else{
+            return productRepository.getBysubCategory(subCategory);
+        }
     }
 
 
     @Override
-    public void deleteProduct(String productName) {
-        productRepository.deleteNode(productName);
+    public void deleteProduct(String productName) throws ProductNotFoundException {
+        if(productRepository.existsById(productName)) {
+            productRepository.deleteNode(productName);
+        }
+        else
+        {
+            throw  new ProductNotFoundException("Products not available");
+
+        }
     }
